+++ test/database.ts
@@ -20,5 +20,5 @@ export const table = new Table({
   name: 'testTable',
   columns: { column1, column2, column3, column4, column5, column6, column7, column8 },
 })
-export const schema = new Schema({ name: 'public', tables: [table] })
+export const schema = new Schema({ name: 'public', tables: { testTable: table }})
 export const database = new Database({ version: 1, schemas: [schema] })